Implementing CombatManager.cs
The player can now charge an animation and use WASD to change what direction the player will be slicing.
when the charge is help of 0.5 seconds and then released the player will play the attack with the last direction selected and then return to idle.

NO CUTTING FUNCTIONALITY IS IMPLEMENTED WITH THE NEW DIRECTIONAL ATTACKS.